Move all simple paths for PyDigraph
